requirements.txt
black==19.10b0
jupyter==1.0.0
pre-commit==2.7.1
pylint==2.4.4
pytest==5.3.5
pytest-icdiff==0.3
coverage==5.2.1
azure-storage-blob==12.3.0
boto3==1.14
google-cloud-storage==1.31.0
sqlalchemy==1.3.18
pandas==1.0.5
model_card_toolkit==0.1.1
